Make decode public, and move it to core
It no longer is coupled to fetcher, and needs to be accessed from outside of that package (by Lens in the immediate-term)
